Revert "Create Collective Order Stats materialized view (#8951)"
This reverts commit d53527c7311bc2021742f08559b04f59eee92511.
