+++ test/utils.js
@@ -50,7 +50,6 @@ export const resetTestDB = async () => {
     await sequelize.query(`TRUNCATE TABLE ${tableNames.join(', ')} RESTART IDENTITY CASCADE`);
     await sequelize.query(`REFRESH MATERIALIZED VIEW "TransactionBalances"`);
     await sequelize.query(`REFRESH MATERIALIZED VIEW "CollectiveBalanceCheckpoint"`);
-    await sequelize.query(`REFRESH MATERIALIZED VIEW "CollectiveOrderStats"`);
   };
 
   try {
